backend/pyproject.toml
[project]
name = "my-personal-examiner-backend"
version = "0.1.0"
description = "PhD-level A-Level Teaching & Examination System - Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework (CONSTITUTIONAL LOCK)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database & ORM (CONSTITUTIONAL LOCK - SQLModel MANDATORY)
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # Configuration
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # AI Integration (Phase III+)
    "openai>=1.54.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    # PDF Processing (Phase II+)
    "pdfplumber>=0.11.0",
    "pypdf>=5.1.0",
    "email-validator>=2.3.0",
    "bcrypt>=4.0.0,<5.0.0",
    # Encryption (Resource Bank - API key encryption)
    "cryptography>=41.0.0",
    # Resource Bank - File Storage (Feature 007)
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "youtube-transcript-api>=0.6.2",
    "boto3>=1.35.0",
    "celery[redis]>=5.4.0",
    "pyclamd>=0.4.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing (CONSTITUTIONAL REQUIREMENT: >80% coverage)
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "moto[s3]>=5.0.0",  # AWS S3 mocking for tests

    # Code Quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",

    # Type Stubs
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
    "types-redis>=4.6.0",
    "boto3-stubs[s3]>=1.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",  # Constitutional requirement
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__init__.py",
    # Phase II models (not implemented yet in Phase I)
    "src/models/subject.py",
    "src/models/syllabus_point.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["alembic.*", "pdfplumber.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
